Extract type into separate file
