+++ src/Microsoft.AspNetCore.OData.Versioning/AspNetCore.Mvc/Versioning/ODataApiVersionActionSelector.cs
@@ -165,46 +165,5 @@
 
             return RoutePolicy.Evaluate( context, selectionResult );
         }
-
-        [DebuggerDisplay( "{Action.DisplayName,nq}" )]
-        sealed class ActionCandidate
-        {
-            internal ActionCandidate( ActionDescriptor action )
-            {
-                Contract.Requires( action != null );
-
-                TotalParameterCount = action.Parameters.Count;
-
-                var filteredParameters = new List<string>( TotalParameterCount );
-
-                for ( var i = 0; i < TotalParameterCount; i++ )
-                {
-                    var parameter = action.Parameters[i];
-
-                    if ( parameter.ParameterType.IsModelBound() )
-                    {
-                        continue;
-                    }
-
-                    var bindingSource = parameter.BindingInfo?.BindingSource;
-
-                    if ( bindingSource != Custom && bindingSource != Path )
-                    {
-                        continue;
-                    }
-
-                    filteredParameters.Add( parameter.Name );
-                }
-
-                Action = action;
-                FilteredParameters = filteredParameters;
-            }
-
-            internal ActionDescriptor Action { get; }
-
-            internal int TotalParameterCount { get; }
-
-            internal IReadOnlyList<string> FilteredParameters { get; }
-        }
     }
 }
